server/requirements.txt
Django==5.1.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
Markdown==3.7
psycopg2-binary==2.9.10